requirement.txt
Django==4.2.4
djangorestframework==3.14.0
ipython==8.15.0
python-genutils==0.2.0
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
django-filter==23.3
